Cast the arguments of the is* functions to int.  Patch from Michael Elkins.

+++ curs_main.c
@@ -536,7 +536,7 @@ void mutt_index_menu (void)
 	    !buf[0])
 	  break;
 
-	if (! isdigit (buf[0]))
+	if (! isdigit ((int) buf[0]))
 	{
 	  mutt_error ("Argument must be a message number.");
 	  break;
